Paths -> PATHS.BIN, path gosub/return
Moved path data into a PATHS.BIN file that is loaded into the game data bank at $A800, and the code is all updated to load that file and use the data at that location.
Implemented gosub/return for paths. max nesting is 2 for now.
Commented out some unused code for now.