More code cleanup, added static entities, fixed PATHS.BIN
Fixed variable naming to be consistent to my style.
Misc code clean up.
Added support for static entities, and placed 32 of them in a row on the bottom at part of the test demo.
Added to and corrected GameData values.
Fixed a bad path in PATHS.BIN, it had the ships facing backwards.