New 32x32 sprites
Changed over to new sprites. Updated most of the code to work with them, still need to fix the hit checks.
Added setting palette offset on sprites, to work with the new sprites that use multiple 16 color palettes.
New sprites are from Kenney.nl
(note, there is a lot of left behind commented stuff for the old sprites that I will remove soon)